Copy static files to S3 when deploying
We have a nasty bug where Cloudfront is caching old files against new
URLs because the new code rolls out gradually across the ~10 admin
instances we have in production.

The way we are going to fix this is by pointing Cloudfront at S3, which
doesn’t have the concept of ‘instances’.

This commit does the work to copy the files to the new buckets. It
depends on the beuckets being set up.