Revert bump to library to unblock functional tests
Version 3.1.3 of govuk-elements-sass makes
everything with the `heading-small` class
`display: block`.

https://github.com/alphagov/govuk_elements/pull/552

We use that class in many places just to make the
text bold with the assumption that it the styles
it applies will not prevent our text rendering
inline. This change breaks that assumption.

We do need to swap these classes out for utility
classes that just apply the bold styles but it can
be done at a later date and we aren't missing out
on much by reverting this library.