Fix duplicated H1 on ‘New letter branding’ page
For accessibility reasons a page should have one (and only one) H1. This
commit fixes an instance where the H1 was duplicated as a result of the
work done to componentize our page headings.

It also adds an extra check to `client_request` so that we don’t
introduce pages with multiple or no H1s in the future.