Update content area blocks in descendent pages
GOV.UK Frontend's template wraps all content in a
`<main>` tag and puts it in the main column.

Some of Notify's pages need to reach out of the
main column (product page).

Some have several blocks of navigation before the
main content area.

This adjusts the blocks in all the templates
where this is the case so those conditions can be
met.